Added is_production variable

+++ mention_action/screenshot_for_blocked.py
@@ -17,11 +17,12 @@ def get_all_links_from_tweet(tweet):
 
 
 class ScreenshotForBlocked(MentionAction):
-    def __init__(self, api):
+    def __init__(self, api, is_production):
         self.api = api
         self.logger = logging.getLogger(__name__)
         self.twitter_status_url = os.environ.get('TWITTER_STATUS_URL', 'https://twitter.com/{}/status/{}')
         self.dark_mode_options = os.environ.get('DARK_MODE_OPTIONS', 'dark').split(',')
+        self.is_production_mode = is_production
 
     async def screenshot_tweet(self, tweet_id, path_to_image, is_dark_mode=False):
         self.logger.debug('Started screenshotting')
@@ -41,28 +42,33 @@ class ScreenshotForBlocked(MentionAction):
     async def reply_to_mention_with_screenshot(self, mention, tweet_to_screenshot_id, add_to_status=''):
         path_to_file = str(tweet_to_screenshot_id) + '.png'
         is_dark_mode = any(dark in mention.text.lower() for dark in self.dark_mode_options)
-        await self.screenshot_tweet(tweet_to_screenshot_id, path_to_file, is_dark_mode)
-        media = self.api.media_upload(path_to_file)
         status = '@' + mention.user.screen_name + ' ' + add_to_status
-        try:
-            self.api.update_status(status=status, in_reply_to_status_id=mention.id,
-                                   media_ids=[media.media_id])
-        except tweepy.TweepError as twe:
-            if twe.api_code == ApiError.RESTRICTED_COMMENTS.value:
-                text = 'נראה שאין לי הרשאות להגיב על הציוץ שביקשת, הנה הציוץ המבוקש'
-                self.logger.info('Cannot comment on mention. sending DM instead of replying')
-                self.api.send_direct_message(recipient_id=mention.user.id, text=text, attachment_type='media',
-                                             attachment_media_id=media.media_id)
+        if self.is_production_mode:
+            await self.screenshot_tweet(tweet_to_screenshot_id, path_to_file, is_dark_mode)
+            media = self.api.media_upload(path_to_file)
+            try:
+                self.api.update_status(status=status, in_reply_to_status_id=mention.id,
+                                       media_ids=[media.media_id])
+            except tweepy.TweepError as twe:
+                if twe.api_code == ApiError.RESTRICTED_COMMENTS.value:
+                    text = 'נראה שאין לי הרשאות להגיב על הציוץ שביקשת, הנה הציוץ המבוקש'
+                    self.logger.info('Cannot comment on mention. sending DM instead of replying')
+                    self.api.send_direct_message(recipient_id=mention.user.id, text=text, attachment_type='media',
+                                                 attachment_media_id=media.media_id)
+                else:
+                    raise twe
             else:
-                raise twe
+                self.logger.info('Reply is successful. path_to_file: {}, status: {}, in_reply_to_status_id: {}, '
+                                 'is_dark_mode: {}'
+                                 .format(path_to_file, status, mention.id, is_dark_mode))
+            finally:
+                if os.path.exists(path_to_file):
+                    self.logger.debug('removing media file')
+                    os.remove(path_to_file)
         else:
             self.logger.info('Reply is successful. path_to_file: {}, status: {}, in_reply_to_status_id: {}, '
                              'is_dark_mode: {}'
                              .format(path_to_file, status, mention.id, is_dark_mode))
-        finally:
-            if os.path.exists(path_to_file):
-                self.logger.debug('removing media file')
-                os.remove(path_to_file)
 
     async def reply_blocked_tweet(self, mention, tweet_id):
         links = ''
@@ -78,8 +84,9 @@ class ScreenshotForBlocked(MentionAction):
         if viewed_tweet.user.id == self.api.me().id:
             msg = 'לגרום לי לעבוד כשהציוץ הוא שלי? אין בושה, הא?'
         self.logger.info(msg)
-        self.api.update_status(status='@' + mention.user.screen_name + ' ' + msg,
-                               in_reply_to_status_id=mention.id)
+        if self.is_production_mode:
+            self.api.update_status(status='@' + mention.user.screen_name + ' ' + msg,
+                                   in_reply_to_status_id=mention.id)
 
     async def blocked_retweet_or_comment(self, mention):
         viewed_tweet = self.api.get_status(mention.in_reply_to_status_id)
@@ -107,8 +114,11 @@ class ScreenshotForBlocked(MentionAction):
                 elif err.api_code == ApiError.NO_TWEET_WITH_ID.value or err.api_code == ApiError.URL_DOESNT_EXIST.value:
                     msg = 'לא הצלחתי למצוא את הציוץ (אולי הוא נמחק?)'
                 if msg != str(err):
-                    self.api.update_status(status='@' + mention.user.screen_name + ' ' + msg,
-                                           in_reply_to_status_id=mention.id)
+                    if self.is_production_mode:
+                        self.api.update_status(status='@' + mention.user.screen_name + ' ' + msg,
+                                               in_reply_to_status_id=mention.id)
+                    else:
+                        pass
                 self.logger.warning(msg)
             except tweepy.TweepError as another_err:
                 self.logger.warning('Unexpected error occurred. error: {}'.format(str(another_err)))
